requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
fvcore==0.1.5.post20221221
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
portalocker==3.2.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.20
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
yacs==0.1.8
zipp==3.23.0
